Fix : lib 버그 해결용, 파일 분리
